modularize all our  backend files in db,routes, models, and controllers directories
