+++ db/index.js
@@ -1,0 +1,4 @@
+import { Sequelize } from "sequelize";
+
+const sequelize = new Sequelize(process.env.PG_URI);
+export default sequelize;
